# api/swagger.yaml
swagger: '2.0'
info:
  description: This is a server for WKON
  version: 1.0.0
  title: WKON
host: 'localhost:5088'
...
securityDefinitions:
  key:
    type: apiKey
    in: header
    name: Key
schemes:
  - http
paths:
...
  /task:
    post:
      security:
        - key: []
      tags:
        - task
      summary: task
      description: This enables a user to create.
      operationId: createtask
      produces:
        - application/xml
        - application/json
      parameters:
        - in: body
          name: body
          description: Created task object
          required: true
          schema:
            $ref: '#/definitions/Task'
      responses:
        default:
          description: successful operation
      x-swagger-router-controller: Task
  '/task/{id}':
    get:
      security:
        - key: []
      tags:
        - task
      summary: Get task by id
      description: ''
      operationId: gettask
      produces:
        - application/xml
        - application/json
      parameters:
        - name: id
          in: path
          description: 'The name that needs to be fetched. Use user1 for testing. '
          required: true
          type: string
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/Task'
        '400':
          description: Invalid Module Id supplied
        '404':
          description: Module not found
      x-swagger-router-controller: Task
    put:
      security:
        - key: []
      tags:
        - task
      summary: Updated task
      description: This can only be done by the logged in user.
      operationId: updatetask
      produces:
        - application/xml
        - application/json
      parameters:
        - name: id
          in: path
          description: name that need to be updated
          required: true
          type: string
        - in: body
          name: body
          description: Updated user object
          required: true
          schema:
            $ref: '#/definitions/Task'
      responses:
        '400':
          description: Invalid user supplied
        '404':
          description: task not found
      x-swagger-router-controller: Task
    delete:
      security:
        - key: []
      tags:
        - task
      summary: Delete module
      description: This can only be done by the logged in user.
      operationId: deletetask
      produces:
        - application/xml
        - application/json
      parameters:
        - name: id
          in: path
          description: The module that needs to be deleted
          required: true
          type: string
      responses:
        '400':
          description: Invalid module supplied
        '404':
          description: module not found
      x-swagger-router-controller: Task
  /task/addassignto:
    put:
      security:
        - key: []
      tags:
        - task
      summary: Updated task
      description: This can only be done by the logged in user.
      operationId: updatetaskmember
      produces:
        - application/xml
        - application/json
      parameters:
        - name: id
          in: query
          description: name that need to be updated
          required: true
          type: string
        - in: body
          name: body
          description: Updated user object
          required: true
          schema:
            $ref: '#/definitions/addMember'
      responses:
        '400':
          description: Invalid user supplied
        '404':
          description: task not found
      x-swagger-router-controller: Task
  /task/deleteassignto:
    put:
      security:
        - key: []
      tags:
        - task
      summary: Updated task
      description: This can only be done by the logged in user.
      operationId: deletetaskByassigntoId
      produces:
        - application/xml
        - application/json
      parameters:
        - name: id
          in: query
          description: name that need to be updated
          required: true
          type: string
        - in: body
          name: body
          description: Updated user object
          required: true
          schema:
            $ref: '#/definitions/addMember'
      responses:
        '400':
          description: Invalid user supplied
        '404':
          description: task not found
      x-swagger-router-controller: Task
  /task/getbysubmoduleid:
    get:
      security:
        - key: []
      tags:
        - task
      summary: Get task by submodule id
      description: ''
      operationId: gettaskBySubmoduleId
      produces:
        - application/xml
        - application/json
      parameters:
        - name: id
          in: query
          description: 'The name that needs to be fetched. Use user1 for testing. '
          required: true
          type: string
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/Task'
        '400':
          description: Invalid Module Id supplied
        '404':
          description: Module not found
      x-swagger-router-controller: Task
  /task/getbyuserId:
    get:
      security:
        - key: []
      tags:
        - task
      summary: Get task by user id
      description: ''
      operationId: gettaskByuserId
      produces:
        - application/xml
        - application/json
      parameters:
        - name: userId
          in: query
          description: The user of the task
          required: true
          type: string
        - name: createdDate
          in: query
          description: The created date for the user
          required: true
          type: string
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/Task'
        '400':
          description: Invalid Module Id supplied
        '404':
          description: Module not found
      x-swagger-router-controller: Task
...
definitions:
...
  Task:
    type: object
    properties:
      name:
        type: string
      description:
        type: string 
      submoduleId:
        type: string
      startDate:
        type: string
      endDate:
        type: string
      date: 
        type: string
      assignTo:
        type: string
      status: 
        type: string    
    example:
      name: name
      description: description
      submoduleId: moduleId
      startDate: startDate
      endDate: endDate
      date: date
      userId: userId
      status: status
    xml:
      name: Task
  addMember:
    type: object
    properties:
      assignTo:
        type: string
    example:
      userId: userId
    xml:
      name: addMember